Improve month filter
Create a new function in SQL that returns the total number of months
(since the year 0). This is then used to query for shows that straddle
the queried month from the frontend.

It is not possible to use date/time manipulation functions as these get
confused when straddling year ends.